+++ src/main/java/com/seb/task/exceptions/HighIncomeException.java
@@ -3,15 +3,16 @@ package com.seb.task.exceptions;
 public class HighIncomeException extends RuntimeException {
    private static final long serialVersionUID = 1L;
 
-   private int income;
+   private double income;
 
-   public HighIncomeException(String message, int income) {
+   public HighIncomeException(String message, double income) {
       super(message);
       this.income = income;
 
    }
 
-   public int getIncome() {
+   public double getIncome() {
       return income;
    }
+
 }
